Fixed CommandLineFile instrumentTest. CommandLineFile: now resets length as well after deleting.

+++ purefm/src/instrumentTest/java/com/docd/purefm/test/CommandLineFileTest.java
@@ -5,8 +5,6 @@ import android.os.Environment;
 import android.test.AndroidTestCase;
 
 import com.docd.purefm.ActivityMonitor;
-import com.docd.purefm.commandline.CommandLine;
-import com.docd.purefm.commandline.ShellHolder;
 import com.docd.purefm.file.CommandLineFile;
 import com.docd.purefm.file.FileFactory;
 import com.docd.purefm.settings.Settings;
@@ -16,7 +14,6 @@ import com.docd.purefm.utils.TextUtil;
 import org.apache.commons.codec.binary.Hex;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -25,7 +22,6 @@ import java.io.IOException;
 public final class CommandLineFileTest extends AndroidTestCase {
 
     private static final File testDir = new File(Environment.getExternalStorageDirectory(), "test");
-    private static final File busybox = new File(Environment.getExternalStorageDirectory(), "busybox");
 
     private static File test1 = new File(testDir, "test1.jpg");
     private static File test2 = new File(testDir, "test2.jpg");
@@ -36,9 +32,6 @@ public final class CommandLineFileTest extends AndroidTestCase {
         final String state = Environment.getExternalStorageState();
         if (!state.equals(Environment.MEDIA_MOUNTED)) {
             throw new RuntimeException("Make sure the external storage is mounted read-write before running this test");
-        }
-        if (!busybox.exists()) {
-            throw new RuntimeException("before running this test copy busybox to " + busybox.getAbsolutePath());
         }
         try {
             FileUtils.forceDelete(testDir);
@@ -55,11 +48,6 @@ public final class CommandLineFileTest extends AndroidTestCase {
         PreviewHolder.initialize(context);
         TextUtil.init(context);
 
-        // override settings to force our test busybox
-        com.docd.purefm.Environment.hasBusybox = true;
-        com.docd.purefm.Environment.busybox = busybox.getAbsolutePath();
-        Settings.useCommandLine = true;
-
         // prepare a test file
         try {
             FileUtils.write(test1, "test");
@@ -70,15 +58,19 @@ public final class CommandLineFileTest extends AndroidTestCase {
 
     @Override
     protected void runTest() {
-        com.docd.purefm.Environment.hasBusybox = true;
+        // override settings to force our test busybox
+        if (!com.docd.purefm.Environment.hasBusybox) {
+            throw new RuntimeException("install busybox on a device before running this test");
+        }
         Settings.useCommandLine = true;
 
-        final CommandLineFile file1 = (CommandLineFile) FileFactory.newFile(test1);
+        CommandLineFile file1 = (CommandLineFile) FileFactory.newFile(test1);
         assertEquals(test1.getAbsolutePath(), file1.getAbsolutePath());
         assertEquals(test1, file1.toFile());
         assertIsNormalFile(file1);
         assertEquals(4, file1.length());
 
+        System.out.println("going to delete");
         file1.delete();
         assertIsEmptyFile(file1);
 
@@ -86,15 +78,18 @@ public final class CommandLineFileTest extends AndroidTestCase {
         assertIsNormalFile(file1);
         final String file1sum = md5sum(file1.toFile());
 
-        final CommandLineFile file2 = (CommandLineFile) FileFactory.newFile(test2);
+        CommandLineFile file2 = (CommandLineFile) FileFactory.newFile(test2);
         assertIsEmptyFile(file2);
 
         file1.move(file2);
         assertIsEmptyFile(file1);
+
+        file2 = (CommandLineFile) FileFactory.newFile(test2);
         assertIsNormalFile(file2);
         assertEquals(file1sum, md5sum(file2.toFile()));
 
         file2.copy(file1);
+        file1 = (CommandLineFile) FileFactory.newFile(test1);
         assertIsNormalFile(file1);
         assertIsNormalFile(file2);
         assertEquals(file1sum, md5sum(file1.toFile()));
